+++ src/ibex_bluesky_core/devices/dae/strategies/__init__.py
@@ -1,16 +1,13 @@
 """Base classes for DAE strategies."""
 
-from ibex_bluesky_core.devices.dae.strategies._base import (
-    Controller,
-    ProvidesExtraReadables,
-    Reducer,
-    Waiter,
-)
+from ibex_bluesky_core.devices.dae import Dae
 from ibex_bluesky_core.devices.dae.strategies._controllers import (
+    Controller,
     PeriodPerPointController,
     RunPerPointController,
     _end_or_abort_run,
 )
+from ibex_bluesky_core.devices.dae.strategies._extra_readables import ProvidesExtraReadables
 from ibex_bluesky_core.devices.dae.strategies._reducers import (
     INTENSITY_PRECISION,
     VARIANCE_ADDITION,
@@ -18,6 +15,7 @@ from ibex_bluesky_core.devices.dae.strategies._reducers import (
     MonitorNormalizer,
     PeriodGoodFramesNormalizer,
     PeriodSpecIntegralsReducer,
+    Reducer,
     ScalarNormalizer,
     sum_spectra,
     tof_bounded_spectra,
@@ -30,12 +28,14 @@ from ibex_bluesky_core.devices.dae.strategies._waiters import (
     PeriodGoodFramesWaiter,
     SimpleWaiter,
     TimeWaiter,
+    Waiter,
 )
 
 __all__ = [
     "INTENSITY_PRECISION",
     "VARIANCE_ADDITION",
     "Controller",
+    "Dae",
     "GoodFramesNormalizer",
     "GoodFramesWaiter",
     "GoodUahWaiter",
